me mata

+++ server/controller/controlador_usuarios.js
@@ -1,17 +1,24 @@
-import { where } from 'sequelize'
 import { User } from '../db.js'
 
-const getUser = async (req, res) => {
-    const receba = await User.findAll()
+const listUser = async (req, res) => {
+    const receba = await User.findAll({ attributes: ['nome','email','status']})
     console.log(receba)
+    res.send(receba)
+}
+
+const getUser = async (req, res) => {
+    const { id } = req.body;
+    const receba = await User.findOne({ where: { id: id }, attributes: ['nome', 'sobrenome', 'email', 'status', 'dataNascimento'] })
     res.send(receba)
 }
 
 const deleteUser = async (req, res) => {
     const { id } = req.body;
     const receba = await User.findOne({ where: { id: id }})
-    res.send(receba)
+    if (!receba) {
+        return res.send('Esse User ai nao existe')
+    }
     const apaga = await User.destroy({where: {id: id}})
 }
 
-export { getUser, deleteUser }
+export { listUser, deleteUser, getUser }
